refactor sql_json view endpoint: move all logic from view to ExecuteSqlCommand (#16682)
